+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## [1.0.1] - 2025-08-07
+### Added
+ - Login page with username/password validation (configurable in config.php)
+ - Logout functionality
+ - Add Validate config
 
 ## [1.0.0] - 2025-08-07
 ### Added